+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 #### Upcoming Changes
 
-*  fix: also mark PC as accessed in run_instruction [#2106](https://github.com/lambdaclass/cairo-vm/pull/2106)
+* feat: implemented delete_unaccessed function [#2099](https://github.com/lambdaclass/cairo-vm/pull/2099)
+
+* fix: also mark PC as accessed in run_instruction [#2106](https://github.com/lambdaclass/cairo-vm/pull/2106)
 
 #### [2.1.0] - 2025-05-21